create RA risk group grid

Running rg_grid() returns a data.table of all 1560 risk groups featured in current RA.

Usage

rg_grid(children = FALSE, kantonal = TRUE)

Arguments

children

Logical with default FALSE. If TRUE age group "0-18" is added to the grid.

kantonal

Logical with default TRUE where Swiss cantons are added to the grid. FALSE returns risk group grid without cantons.

Value

A data.table containing all combinations of

Kanton

26 swiss cantons

Altersklasse

15 age groups (16 when children = T)

Geschlecht

M/F

Aufenthalt

J/N dummy variable indicating whether person was hospitalized for at least 3 consecutive nights

See also

Author

Magnus Vieten, Yannick Schwarz

Examples

rg_grid()
#>       Kanton Altersklasse Geschlecht Aufenthalt
#>       <char>       <char>     <char>     <char>
#>    1:     AG        19-25          F          J
#>    2:     AG        19-25          F          N
#>    3:     AG        19-25          M          J
#>    4:     AG        19-25          M          N
#>    5:     AG        26-30          F          J
#>   ---                                          
#> 1556:     ZH        86-90          M          N
#> 1557:     ZH          91+          F          J
#> 1558:     ZH          91+          F          N
#> 1559:     ZH          91+          M          J
#> 1560:     ZH          91+          M          N