Teuerung

Calculates the cost increase ('Niveauteuerung pro Kanton') used as an extrapolation factor in risk equalization. Derivation see: Berechnungsformeln fuer den Risikoasugleich mit PCG ab 2020, Section B.

Usage

teuerung_fun(dt_T_14, dt_T_1_14)

Arguments

dt_T_14

A data.table featuring individual insurance data according to naming and data type standards used in data set test_data.

dt_T_1_14

A data.table featuring individual insurance data according to naming and data type standards used in data set test_data.

Value

A data.table containing

Kanton

Swiss canton

pi_Niv_k

Total cost increase component divided by structural cost increase component ('Niveateuerung')

pi_Str_k1

Structural cost increase component

pi_Ges_k1

Total cost increase component

See also

Author

Magnus Vieten, Yannick Schwarz

Examples

if (FALSE) { # \dontrun{
pcg_grouping(data_input = some_RA_IndData_object, pcg_liste = "path/to/prepped/pcg_list.xlsx")
} # }